requirements.txt
fastapi
uvicorn
python-multipart
streamlit
rembg
pillow
onnxruntime
numpy==1.26.4